ref(k8s): extract consumer builder from pipeline_step
Move shared consumer deployment logic into consumer_builder so it can
be reused by the pipeline macro and the streaming operator.
